As discussed, flatten the oauth code The OAuth/OAuthHeaderGenerator was only used in one place And copies data into OAuthHeaderData in order to then copy it into the crenna OAuthRequest Testing the results of oauth header generation is almost as easy as the RequestBuilder now just returns data so the tests are still there in RequestBuilderOAuthHeaderTests
