+++ src/SevenDigital.Api.Wrapper/SevenDigital.Api.Wrapper.csproj
@@ -49,8 +49,6 @@
     <Compile Include="AttributeManagement\AttributeRequestDataBuilder.cs" />
     <Compile Include="EndpointResolution\AppDomainAssemblyResolver.cs" />
     <Compile Include="EndpointResolution\EssentialDependencyCheck.cs" />
-    <Compile Include="EndpointResolution\OAuth\OAuthHeaderData.cs" />
-    <Compile Include="EndpointResolution\OAuth\OAuthHeaderGenerator.cs" />
     <Compile Include="EndpointResolution\RequestHandlers\IRequestBuilder.cs" />
     <Compile Include="EndpointResolution\RequestHandlers\RequestBuilder.cs" />
     <Compile Include="EndpointResolution\RequestHandlers\RouteParamsSubstitutor.cs" />
